# nodejs-client/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Rakam API Documentation",
    "description": "An analytics platform API that lets you create your own analytics services.",
    "version": "1.0.0"
  },
  "produces": ["application/json"],
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
...
  "/recipe/export": {
    
    "post": {
      "summary": "Export recipe",
      "description":"",
      "x-swagger-router-controller": "Recipe",
      "tags": ["Recipe"],
      "operationId": "export",
      "parameters": [
        {
  "name" : "project",
  "in" : "formData",
  "required" : false,
  "type" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "successful operation",
  "schema" : {
    "$ref" : "#/definitions/Recipe"
  }
}
        
        
      }
    } 
    
  } ,
...
  "/project/create": {
    
    "post": {
      "summary": "Create project",
      "description":"",
      "x-swagger-router-controller": "Admin",
      "tags": ["Admin"],
      "operationId": "createProject",
      "parameters": [
        {
  "name" : "name",
  "in" : "formData",
  "required" : false,
  "type" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "successful operation",
  "schema" : {
    "$ref" : "#/definitions/JsonResponse"
  }
}
        
        
      }
    } 
    
  } ,
...
  "/project/schema": {
    
    "post": {
      "summary": "Get collection schema",
      "description":"",
      "x-swagger-router-controller": "Admin",
      "tags": ["Admin"],
      "operationId": "schema",
      "parameters": [
        {
  "name" : "project",
  "in" : "formData",
  "required" : false,
  "type" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "successful operation",
  "schema" : {
    "type" : "array",
    "items" : {
      "$ref" : "#/definitions/Collection"
    }
  }
}
        ,
        "400": {
  "description" : "Project does not exist."
}
        
        
      }
    } 
    
  } ,
...
  }, "definitions": {
  "ContinuousQueryBuilder": {
  "properties" : {
    "name" : {
      "type" : "string"
    },
    "query" : {
      "type" : "string"
    },
    "collections" : {
      "type" : "array",
      "items" : {
        "type" : "string"
      }
    },
    "options" : {
      "type" : "object",
      "additionalProperties" : {
        "type" : "object",
        "properties" : { }
      }
    },
    "tableName" : {
      "type" : "string"
    },
    "partitionKeys" : {
      "type" : "array",
      "items" : {
        "type" : "string"
      }
    }
  }
},"User": {
...
},"Recipe": {
  "required" : [ "collections", "reports", "strategy" ],
  "properties" : {
    "strategy" : {
      "type" : "string",
      "enum" : [ "DEFAULT", "SPECIFIC" ]
    },
    "project" : {
      "type" : "string"
    },
    "collections" : {
      "type" : "object",
      "additionalProperties" : {
        "$ref" : "#/definitions/Collection"
      }
    },
    "reports" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/ReportBuilder"
      }
    },
    "materializedViews" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/MaterializedViewBuilder"
      }
    },
    "continuousQueries" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/ContinuousQueryBuilder"
      }
    },
    "continuousQueryBuilders" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/ContinuousQueryBuilder"
      }
    },
    "materializedViewBuilders" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/MaterializedViewBuilder"
      }
    }
  }
},"BatchCreate": {
...
},"JsonResponse": {
  "properties" : {
    "success" : {
      "type" : "boolean",
      "default" : false
    },
    "message" : {
      "type" : "string"
    }
  }
},"Send": {
...
},"ReportBuilder": {
  "properties" : {
    "slug" : {
      "type" : "string"
    },
    "name" : {
      "type" : "string"
    },
    "query" : {
      "type" : "string"
    },
    "options" : {
      "type" : "object",
      "additionalProperties" : {
        "type" : "object",
        "properties" : { }
      }
    }
  }
},"Measure": {
...
},"MaterializedViewBuilder": {
  "properties" : {
    "name" : {
      "type" : "string"
    },
    "table_name" : {
      "type" : "string"
    },
    "query" : {
      "type" : "string"
    },
    "options" : {
      "type" : "object",
      "additionalProperties" : {
        "type" : "object",
        "properties" : { }
      }
    },
    "updateInterval" : {
      "type" : "string"
    }
  }
},"ABTestingReport": {
...
},"SchemaField": {
  "properties" : {
    "name" : {
      "type" : "string",
      "readOnly" : true
    },
    "type" : {
      "type" : "string",
      "readOnly" : true,
      "enum" : [ "STRING", "LONG", "DOUBLE", "BOOLEAN", "DATE", "TIME", "TIMESTAMP", "ARRAY_STRING", "ARRAY_LONG", "ARRAY_DOUBLE", "ARRAY_BOOLEAN", "ARRAY_DATE", "ARRAY_TIME", "ARRAY_TIMESTAMP" ]
    },
    "nullable" : {
      "type" : "boolean",
      "readOnly" : true,
      "default" : false
    },
    "unique" : {
      "type" : "boolean",
      "readOnly" : true,
      "default" : false
    },
    "descriptiveName" : {
      "type" : "string",
      "readOnly" : true
    },
    "description" : {
      "type" : "string",
      "readOnly" : true
    },
    "category" : {
      "type" : "string",
      "readOnly" : true
    }
  }
},"AutomationRule": {
...
},"Collection": {
  "properties" : {
    "name" : {
      "type" : "string"
    },
    "fields" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/SchemaField"
      }
    }
  }
},"Event": {
...
  }
}